feat: fixes after testing with noisier data (#2)
* print a warning to console when invalid coordinates are present in dataset.
* rename cli to __main__ to allow calling as a python module
* move data validation to `io` module

+++ geocluster/clustering.py
@@ -50,5 +50,5 @@ def cluster_locations(df, algorithm, radius_km, min_cluster_size):
             n_jobs=-1,
         )
 
-    X = np.radians(np.array(coordinates))
+    X = np.radians(np.array(coordinates).astype(float))
     return to_cluster_dict(df, clustering.fit(X))
